Update font path in src/fonts.css
The paths for Inter & WorkSans font files were updated. Previously, they were directed to an incorrect folder (../fonts/) causing the fonts to not be displayed properly. This change points to the correct folder so that the site correctly uses the intended fonts in its design. This will greatly enhance the readability and overall aesthetics of the website.